creating sensor managers
