+++ SlimHUD/Services/AppleScriptRunner.swift
@@ -1,0 +1,20 @@
+//
+//  AppleScript.swift
+//  MusicBar
+//
+//  Created by Alex Perathoner on 26/12/2019.
+//
+import Foundation
+
+class AppleScriptRunner {
+    private init() {}
+    
+    static func run(script: String) -> String? {
+        var error: NSDictionary?
+        if let scriptObject = NSAppleScript(source: script) {
+            let output: NSAppleEventDescriptor = scriptObject.executeAndReturnError(&error)
+            return output.stringValue
+        }
+        return nil
+    }
+}
