+++ SlimHUD/Services/BrightnessManager.swift
@@ -1,0 +1,37 @@
+//
+//  DisplayManager.swift
+//  SlimHUD
+//
+//  Created by Alex Perathoner on 19/12/2022.
+//  Copyright © 2022 Alex Perathoner. All rights reserved.
+//
+
+import Foundation
+
+class BrightnessManager {
+    private init() {}
+    
+    static func getDisplayBrightness() -> Float {
+        var brightness: float_t = 1
+        let service = IOServiceGetMatchingService(kIOMasterPortDefault, IOServiceMatching("IODisplayConnect"))
+        
+        IODisplayGetFloatParameter(service, 0, kIODisplayBrightnessKey as CFString, &brightness)
+        IOObjectRelease(service)
+        return brightness
+    }
+    
+    static func getKeyboardBrightness() -> Float {
+        let service = IOServiceGetMatchingService(kIOMasterPortDefault, IOServiceMatching("AppleHIDKeyboardEventDriverV2"))
+        defer {
+            IOObjectRelease(service)
+        }
+        
+        if let ser: CFTypeRef = IORegistryEntryCreateCFProperty(service, "KeyboardBacklightBrightness" as CFString, kCFAllocatorDefault, 0)?.takeUnretainedValue() {
+            let result = ser as! Float
+            return result / 342 //max value is 342, proportioning to %
+        }
+        //couldn't get keyboard backlight
+        return 0.5
+    }
+
+}
